fix(condo): DOMA-6070 move helpers functions with remove recurrent payment functionality into separate tasks
