+++ docs/FAQ.md
@@ -14,7 +14,7 @@ Require stack:
     at Function.Module._resolveFilename (internal/modules/cjs/loader.js:980:15)
     at Function.Module._load (internal/modules/cjs/loader.js:862:27)
     at Module.require (internal/modules/cjs/loader.js:1042:19)
-    at require (internal/modules/cjs/helpers.js:77:18)
+    at require (internal/modules/cjs/removeRecurrentPayments.js:77:18)
     at Object.<anonymous> (/Users/pahaz/Code/pahaz/my-new-startup/bin/createapp:3:23)
     at Module._compile (internal/modules/cjs/loader.js:1156:30)
     at Object.Module._extensions..js (internal/modules/cjs/loader.js:1176:10)
